+++ mpeg2dec/doc/sample2.c
@@ -69,7 +69,7 @@ static void sample2 (FILE * mpgfile)
     }
     info = mpeg2_info (decoder);
 
-    size = -1;
+    size = (size_t)-1;
     do {
 	state = mpeg2_parse (decoder);
 	switch (state) {
